fix: split EmojiResolver
